kernel: fixed compilation issues
